Added conversion for ultrasonic.

+++ FRC4536-2016-Robot/src/org/usfirst/frc/team4536/robot/Constants.java
@@ -31,7 +31,7 @@ public class Constants {
 	public static final double HOLD_SPEED = 0.00;
 	public static final double INTAKE_STICTION = 0.06; //Full battery
 	
-	public static final int MAX_ULTRA_CONVERSION = 92;
+	public static final double MAX_ULTRA_CONVERSION = 92.0;
 	
 	/*-----Access Values-----*/
 	/*@author Liam
